pyproject.toml
[tool.poetry]
name = "eff-mobile"
version = "0.1.0"
description = ""
authors = ["anatoliy.klimanov@gmail.com"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.4"
django-filter = "^24.3"
psycopg2-binary = "^2.9.10"
isort = "^5.13.2"
ruff = "^0.9.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 80
exclude = [
  "migrations",
  "settings.py",
  "apps.py",
  "restaurant/orders/tests.py",
  "restaurant/orders/models.py",

]

[tool.ruff.lint]
preview = true
select = ["E", "F", "I", "C90"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]